refactor: estruturando as pastas do front
