+++ frontend/frontend-app/src/views/dashboard/DashboardView.vue
